+++ packages/babel-plugin-jsx-sfc/index.d.ts
@@ -1,5 +1,5 @@
 /*!
- * babel-plugin-jsx-sfc v1.0.2
+ * babel-plugin-jsx-sfc v1.1.0
  * (c) 2020-present Joe_Sky
  * Released under the MIT License.
  */